rest_open_course_appointments/requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.25
Django==3.1.4
django-celery==3.3.1
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
kombu==3.0.37
Markdown==3.3.3
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
